requirements.txt
flask==3.1.0
httpx==0.28.1
pandas==2.2.3
python-dotenv==1.1.0